requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.1.0
attrs==23.1.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distlib==0.3.8
distro==1.8.0
extra-streamlit-components==0.1.60
filelock==3.13.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
identify==2.5.33
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langsmith==0.0.71
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.2
openai==1.4.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pip==23.2.1
platformdirs==4.1.0
pre-commit==3.6.0
protobuf==4.25.1
pyarrow==14.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
setuptools==68.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
st-pages==0.4.5
streamlit==1.29.0
streamlit-authenticator==0.2.3
streamlit-custom-notification-box==0.1.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
virtualenv==20.25.0
watchdog==3.0.0
wheel==0.41.2
yarl==1.9.4
zipp==3.17.0
